pyproject.toml
[project]
name = "ngtrader"
version = "0.1.0"
description = "Review market price action, monitor portfolio strategies, and execute live trades with AI"
readme = "README.md"
requires-python = ">=3.12"
license = "BSD-3-Clause"
dependencies = [
    "ib-async>=2.1.0",
    "langgraph>=0.2.76",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pandera[pandas]>=0.20",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
